add upload file crud for company and logic for file uploading to cloudinary

+++ api/app/company/crud.py
@@ -1,17 +1,17 @@
 from typing import List
 
-from fastapi import HTTPException, status
+from fastapi import HTTPException, status, UploadFile
 from sqlmodel import select
 
 from api.app.common.dependencies import SessionDep
 from api.app.company.models import Company
 from api.app.company.schemas import CompanyCreate, CompanyResponse, CompanyUpdate
-
+from api.app.utils import upload_file, delete_from_cloudinary
 
 COMPANY_NOT_FOUND = "Company not found"
 
 
-def create_company(session: SessionDep, company: CompanyCreate) -> CompanyResponse:
+async def create_company(session: SessionDep, company: CompanyCreate, image: UploadFile) -> CompanyResponse:
     existing_company = session.exec(
         select(Company).filter(Company.title == company.title)
     ).first()
@@ -21,10 +21,13 @@ def create_company(session: SessionDep, company: CompanyCreate) -> CompanyRespon
             status_code=status.HTTP_400_BAD_REQUEST, detail="Company already exists"
         )
 
+    image_data: dict = await upload_file(image)
+
     db_company = Company(
         title=company.title,
         description=company.description,
-        image_link=company.image_link,
+        image_link=image_data.get('url'),
+        image_id=image_data.get('image_id'),
         country_id=company.country_id,
         kitchen_id=company.kitchen_id,
     )
@@ -37,7 +40,7 @@ def create_company(session: SessionDep, company: CompanyCreate) -> CompanyRespon
 
 
 def get_all_companies(
-    session: SessionDep, page: int = 1, limit: int = 10
+        session: SessionDep, page: int = 1, limit: int = 10
 ) -> List[CompanyResponse]:
     return session.exec(select(Company).limit(limit).offset((page - 1) * limit)).all()
 
@@ -54,7 +57,7 @@ def get_company_by_id(session: SessionDep, company_id: int) -> CompanyResponse:
 
 
 def update_company(
-    session: SessionDep, company: CompanyUpdate, company_id: int
+        session: SessionDep, company: CompanyUpdate, company_id: int
 ) -> CompanyResponse:
     existing_company: Company = session.exec(
         select(Company).where(Company.id == company_id)
@@ -76,7 +79,7 @@ def update_company(
 
 
 def remove_company(session: SessionDep, company_id: int) -> CompanyResponse:
-    existing_company = session.exec(
+    existing_company: Company = session.exec(
         select(Company).where(Company.id == company_id)
     ).first()
 
@@ -84,6 +87,11 @@ def remove_company(session: SessionDep, company_id: int) -> CompanyResponse:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail=COMPANY_NOT_FOUND
         )
+
+    result = delete_from_cloudinary(existing_company.image_id)
+
+    if not result:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="The image could not be deleted")
 
     session.delete(existing_company)
     session.commit()
